Aggiunto design pattern memento.

+++ src/main/java/it/unicam/cs/CityTourNet/handlers/CompraVenditaHandler.java
@@ -1,14 +1,18 @@
 package it.unicam.cs.CityTourNet.handlers;
 
 import it.unicam.cs.CityTourNet.model.Notifica;
+import it.unicam.cs.CityTourNet.model.contenuto.Contenuto;
+import it.unicam.cs.CityTourNet.model.contenuto.ContenutoMemento;
 import it.unicam.cs.CityTourNet.model.contenuto.ProdottoGadget;
 import it.unicam.cs.CityTourNet.model.utente.*;
+import it.unicam.cs.CityTourNet.repositories.ContenutoMementoRepository;
 import it.unicam.cs.CityTourNet.repositories.ContenutoRepository;
 import it.unicam.cs.CityTourNet.repositories.NotificaRepository;
 import it.unicam.cs.CityTourNet.repositories.UtenteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Comparator;
 import java.util.List;
 
 
@@ -18,13 +22,16 @@ public class CompraVenditaHandler {
     private final ContenutoRepository contenutoRepository;
     private final UtenteRepository utenteRepository;
     private final NotificaRepository notificaRepository;
+    private final ContenutoMementoRepository contenutoMementoRepository;
 
     @Autowired
-    public CompraVenditaHandler(ContenutoRepository contenutoRepository,
-                                UtenteRepository utenteRepository, NotificaRepository notificaRepository) {
+    public CompraVenditaHandler(ContenutoRepository contenutoRepository, UtenteRepository utenteRepository,
+                                NotificaRepository notificaRepository,
+                                ContenutoMementoRepository contenutoMementoRepository) {
         this.contenutoRepository = contenutoRepository;
         this.utenteRepository = utenteRepository;
         this.notificaRepository = notificaRepository;
+        this.contenutoMementoRepository = contenutoMementoRepository;
     }
 
     public ProdottoGadget getProdottoGadget(long id){
@@ -37,19 +44,27 @@ public class CompraVenditaHandler {
     public List<ProdottoGadget> getProdottiGadget(){
         return this.contenutoRepository.findAll()
                 .stream()
-                .filter(c -> c.getTipoContenuto().equals("ProdottoGadget"))
+                .filter(c -> c instanceof ProdottoGadget)
                 .map(c -> (ProdottoGadget) c)
                 .toList();
     }
 
-    public boolean addProdottoGadget(ProdottoGadget gadget) {
+    public void addProdottoGadget(ProdottoGadget gadget) {
+        gadget.setDefinitive(true);
         this.contenutoRepository.saveAndFlush(gadget);
-        return true;
+        this.salvaStato(gadget);
     }
 
-    public boolean removeProdottoGadget(long id){
-        this.contenutoRepository.deleteById(id);
-        return true;
+    public void removeProdottoGadget(long id){
+        if(this.contenutoRepository.existsById(id)) {
+            Contenuto daEliminare = this.contenutoRepository.findById(id).get();
+            List<ContenutoMemento> mementoStack = this.contenutoMementoRepository.findAll()
+                    .stream()
+                    .filter(m -> m.getIDContenuto() == daEliminare.getID())
+                    .toList();
+            this.contenutoMementoRepository.deleteAll(mementoStack);
+            this.contenutoRepository.deleteById(id);
+        }
     }
 
     public int getPuntiUtente(String username){
@@ -60,11 +75,10 @@ public class CompraVenditaHandler {
         return -1;
     }
 
-    private boolean riduciNumeroPezziDisponibili(long id, int numPezzi) {
+    private void riduciNumeroPezziDisponibili(long id, int numPezzi) {
         ProdottoGadget daAcquistare = this.getProdottoGadget(id);
         daAcquistare.setNumPezzi(daAcquistare.getNumPezzi() - numPezzi);
         this.contenutoRepository.saveAndFlush(daAcquistare);
-        return true;
     }
 
     public boolean gestisciAcquistoProdottoGadget(long id, int numPezzi,
@@ -90,32 +104,67 @@ public class CompraVenditaHandler {
     }
 
 
-    private boolean inviaNotificaAcquistoConfermato(String usernameAutore,
-                                                    String usernameAcquirente,
-                                                    String nomeProdottoGadget, int numPezzi,
-                                                    int prezzo, String indirizzo){
+    private void inviaNotificaAcquistoConfermato(String usernameAutore,
+                                                 String usernameAcquirente,
+                                                 String nomeProdottoGadget, int numPezzi,
+                                                 int prezzo, String indirizzo){
         String testo = "Hai acquistato " + numPezzi + " pezzi del prodotto: " + nomeProdottoGadget + ".\n"
                 + "Il costo totale e': " + prezzo + "\n" +
                 "Il prodotto verra' spedito a breve a questo indirizzo: " + indirizzo;
         Notifica notifica = new Notifica(usernameAutore, usernameAcquirente, testo);
         this.notificaRepository.saveAndFlush(notifica);
-        return true;
     }
 
-    private boolean inviaNotificaPezziInsufficienti(String usernameAutore,
-                                                    String usernameAcquirente) {
+    private void inviaNotificaPezziInsufficienti(String usernameAutore,
+                                                 String usernameAcquirente) {
         String testo = "Il numero di pezzi selezionato non è disponibile";
         Notifica notifica = new Notifica(usernameAutore, usernameAcquirente, testo);
         this.notificaRepository.saveAndFlush(notifica);
-        return true;
     }
 
-    private boolean inviaNotificaPuntiInsufficienti(String usernameAutore,
-                                                    String usernameAcquirente) {
+    private void inviaNotificaPuntiInsufficienti(String usernameAutore,
+                                                 String usernameAcquirente) {
         String testo = "Non possiedi abbastanza punti per effettuare l'acquisto";
         Notifica notifica = new Notifica(usernameAutore, usernameAcquirente, testo);
         this.notificaRepository.saveAndFlush(notifica);
-        return true;
+    }
+
+    private void salvaStato(Contenuto contenuto){
+        this.contenutoMementoRepository.saveAndFlush(contenuto.createMemento());
+    }
+
+    public boolean recuperaStato(Contenuto contenuto) {
+        List<ContenutoMemento> mementoStack = this.contenutoMementoRepository.findAll()
+                .stream().filter(m -> m.getIDContenuto() == contenuto.getID())
+                .sorted(Comparator.comparingLong(ContenutoMemento::getId))
+                .toList();
+        if (mementoStack.size() > 1) {
+            ContenutoMemento daRecuperare = mementoStack.get(mementoStack.size() - 2);
+            ContenutoMemento daEliminare = mementoStack.get(mementoStack.size() - 1);
+            contenuto.restoreMemento(daRecuperare);
+            this.contenutoMementoRepository.deleteById(daEliminare.getId());
+            return true;
+        }
+        return false;
+    }
+
+    public boolean annullaModifiche(long id) {
+        if(this.contenutoRepository.existsById(id)) {
+            return this.recuperaStato(this.contenutoRepository.findById(id).get());
+        }
+        return false;
+    }
+
+    public boolean eseguiModifiche(int prezzo, int numPezzi, long id) {
+        if(this.contenutoRepository.existsById(id)) {
+            if(this.contenutoRepository.findById(id).get() instanceof ProdottoGadget prodottoGadget) {
+                prodottoGadget.setPrezzo(prezzo);
+                prodottoGadget.setNumPezzi(numPezzi);
+                this.salvaStato(prodottoGadget);
+                return true;
+            }
+        }
+        return false;
     }
 
 }
